% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chg_rownames.R
\name{chg_rownames}
\alias{chg_rownames}
\title{Changes rownames of a matrix}
\usage{
chg_rownames(matrix, newnames)
}
\arguments{
\item{matrix}{any matrix of dimension n*m}

\item{newnames}{a character vector of length n}
}
\value{
the given matrix with changed rownames
}
\description{
Function takes a vector containing names and a matrix and exchanges
the rownames of the matrix with the ones provided in the names vector.
Note that the row dimension of the matrix has to be equal to the length of the
names vector. The function was written to enable a rownames replacement in diplyr's pipe.
}
