% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gene_names.R
\name{extract_gene_names}
\alias{extract_gene_names}
\title{Creates a ordered list of colnames of a given Matrix}
\usage{
extract_gene_names(matrix)
}
\arguments{
\item{matrix}{a numerical matrix}
}
\value{
vector containing colnames
}
\description{
This function orders the columns of a matrix by variance starting with the highest within column variance
left to the lowest within column variance on the right.
}
\examples{
\dontrun{
colnames <- matrix \%>\%
  extract_gene_names()
}
}
