% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Creates gene expression matrix and gene list from the Golub data}
\usage{
fetch_data()
}
\value{
loads expression data and a vector with gene names into GlobalEnv
}
\description{
Calling this function will load the 'golubEsets' and 'tidyverse' libraries.
Furthermore, the Golub_Train matrix with gene espression data will be loaded and transposed.
After transposing it, zeros will be replaced by ones and the expression values are normalised
by applying log2. Afterwards, rownames of the expression matrix wel be changed from patient IDs
to the corresponding cancertypes. Beside the expression data, a character vector is created
which contains gene names in order of their highest to lowest expression value.
}
