% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_expr_data.R
\name{load_expr_data}
\alias{load_expr_data}
\title{Creates gene expression matrix from the Golub data}
\usage{
load_expr_data()
}
\value{
normalized expression matrix of Golub data
}
\description{
Calling this function will load the 'golubEsets' and 'tidyverse' libraries.
Furthermore, the Golub_Train matrix with gene espression data will be loaded and transposed.
After transposing it, zeros will be replaced by ones and the expression values are normalised
by applying log2. Afterwards, rownames of the expression matrix are changed from patient IDs
to the corresponding cancertypes.
}
