% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_zeros.R
\name{replace_zeros}
\alias{replace_zeros}
\title{Replaces zeros in a numerical matrix}
\usage{
replace_zeros(matrix, x)
}
\arguments{
\item{matrix}{numerical matrix}

\item{x}{numerical value to replace zeros with}
}
\value{

}
\description{
Replaces zeros in a numerical matrix with a given numerical value x. Was written to enable
the matrix operator to be used within a diplyr pipe
}
