% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runapp}
\alias{runapp}
\title{ShinyAPP displaing hierarchical clustering and PCA of gene expression data}
\usage{
runapp()
}
\value{
starts a shiny App to be displayed inside the webbrowser
}
\description{
ShinyAPP displaing hierarchical clustering and PCA of gene expression data
}
